changed message type to byte length

+++ Constants.java
@@ -3,14 +3,14 @@ public class Constants {
 		public static String PEER_CFG_FILE  = "PeerInfo.cfg";
                 public static String COMMON_CFG_FILE = "Common.cfg";
 		public static String PDIR = "./";
-		public static int MSG_CHOKE_TYPE = 0;
-		public static int MSG_UNCOKE_TYPE = 1;
-		public static int MSG_INTERESTED_TYPE = 2;
-		public static int MSG_UNINTERESTED_TYPE = 3;
-		public static int MSG_HAVE_TYPE = 4;
-		public static int MSG_BITFIELD_TYPE = 5;
-		public static int MSG_REQUEST_TYPE = 6;
-		public static int MSG_PIECE_TYPE = 7;
+		public static byte MSG_CHOKE_TYPE = 0;
+		public static byte MSG_UNCOKE_TYPE = 1;
+		public static byte MSG_INTERESTED_TYPE = 2;
+		public static byte MSG_UNINTERESTED_TYPE = 3;
+		public static byte MSG_HAVE_TYPE = 4;
+		public static byte MSG_BITFIELD_TYPE = 5;
+		public static byte MSG_REQUEST_TYPE = 6;
+		public static byte MSG_PIECE_TYPE = 7;
 		public static String HANDSHAKE_HEADER = "HELLO";
 		public static final long HANDSHAKE_UID = 3L;
